Added factory protocol -> abstraction, composition, debugging
